changed timer cost from double to int

+++ ddl.sql
@@ -85,7 +85,7 @@ CREATE TABLE timer (
   pool_id BIGINT REFERENCES pool(id),
   installed INT NOT NULL,
   model VARCHAR NOT NULL,
-  cost NUMERIC(7, 2) NOT NULL
+  cost INT NOT NULL
 );
 
 CREATE TABLE timer_setting (
