+++ shared/src/main/scala/poolmate/Entity.scala
@@ -126,7 +126,7 @@ final case class Timer(id: Long = 0,
                        poolId: Long = 0,
                        installed: Int = 0,
                        model: String = "",
-                       cost: Double = 0.0) extends Entity:
+                       cost: Int = 0) extends Entity:
   def display = model
 
 final case class TimerSetting(id: Long = 0,
